Silenced several compilation warnings

+++ unsupported/Eigen/CXX11/src/Tensor/TensorAssign.h
@@ -103,7 +103,7 @@ struct TensorEvaluator<const TensorAssignOp<LeftArgType, RightArgType>, Device>
       m_leftImpl(op.lhsExpression(), device),
       m_rightImpl(op.rhsExpression(), device)
   {
-    EIGEN_STATIC_ASSERT((TensorEvaluator<LeftArgType, Device>::Layout == TensorEvaluator<RightArgType, Device>::Layout), YOU_MADE_A_PROGRAMMING_MISTAKE);
+    EIGEN_STATIC_ASSERT((static_cast<int>(TensorEvaluator<LeftArgType, Device>::Layout) == static_cast<int>(TensorEvaluator<RightArgType, Device>::Layout)), YOU_MADE_A_PROGRAMMING_MISTAKE);
     // The dimensions of the lhs and the rhs tensors should be equal to prevent
     // overflows and ensure the result is fully initialized.
     eigen_assert(dimensions_match(m_leftImpl.dimensions(), m_leftImpl.dimensions()));
